mcp-servers/Wasabi/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
sse-starlette>=2.0.0
mcp>=0.1.0
fastmcp==2.9.0
boto3==1.38.46
pypdf
python-docx
pandas
xlrd
openpyxl